+++ src/main/java/_1danhebojo/coalarm/coalarm_service/domain/alert/service/AlertHistoryService.java
@@ -16,6 +16,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Instant;
@@ -29,6 +30,7 @@ public class AlertHistoryService {
 
     private final AlertHistoryRepositoryImpl alertHistoryRepository;
 
+    // 알람 리스트 조회
     @Transactional(readOnly = true)
     public AlertHistoryListResponse getAlertHistoryList(Long userId, PaginationRequest paginationRequest) {
         int offset = paginationRequest.getOffset();
@@ -49,6 +51,7 @@ public class AlertHistoryService {
         );
     }
 
+    // 알람 정보 조회
     public AlertHistoryResponse getAlertHistory(Long alertHistoryId) {
         AlertHistory alertHistory = alertHistoryRepository.findById(alertHistoryId)
                 .orElseThrow(() -> new RuntimeException("해당 알람 히스토리를 찾을 수 없습니다."));
@@ -56,7 +59,8 @@ public class AlertHistoryService {
         return new AlertHistoryResponse(alertHistory);
     }
 
-    @Transactional
+    // 알람 히스토리 저장
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void addAlertHistory(Long alertId, Long userId) {
         AlertHistory alertHistory = new AlertHistory();